JavaScript - Dynamic HTML assembly function & enum
This update adds the main DOM element assembler `assembleElement`, which will be called in bulk by a yet-undefined handler function to dynamically build the interface once the DOM is rendered and `init` invoked. The module will then employ the fade-in function to fade-in on the page, jQuery style.

Additionally, adds an enum specifically dedicated to the storage of dynamic element ids and class names to prevent unnecessary duplicates appearing the module body. This makes any adjustment between the JS and CSS files much easier to undertake in the event of an identifier signature change.